Cookie creation on login and backend auth

+++ controllers/paymentController.js
@@ -3,17 +3,20 @@ var paymentModel = require('../models/payment.js');
 module.exports = {
    
     viewAllForUser: async function (app, req, res) {
+        if(!req.session.login){return res.status(401).json({error: 'Unauthorised'})}
         res = await paymentModel.getPaymentsForUser(app, req.session.userId, res)
         return res;
     },
 
     addItem: async function(app, req, res){
+        if(!req.session.login){return res.status(401).json({error: 'Unauthorised'})}
         let payment = paymentModel.createPaymentFromRequest(req);
         res = await paymentModel.addPayment(app, payment, res)
         return res;
     }, 
     
     amendItem: async function (app, req, res) {
+        if(!req.session.login){return res.status(401).json({error: 'Unauthorised'})}
         let amendedPayment = paymentModel.createPaymentFromRequest(req);
         res = await updatePayment(app, amendedPayment, req.body.paymentId, res)
         return res;
@@ -22,6 +25,7 @@ module.exports = {
     deleteItem: async function (app, req, res) {
         // console.log(req.body)
         // return res.status(400).json({error: 'Request Body: ' + req.body.paymentId})
+        if(!req.session.login){return res.status(401).json({error: 'Unauthorised'})}
         res = await paymentModel.deletePayment(app, req.body.paymentId, res);
         return res;
     },
