+++ controllers/userController.js
@@ -18,22 +18,25 @@ module.exports = {
 
     addItem: async function (app, req, res) {   
         let newUser =  await userModel.createUserFromRequest(req);                                
-        res = await userModel.addUser(app,newUser, res);
+        res = await userModel.addUser(app, newUser, req.session, res);
         return res;
     },
 
     getItem: async function (app, req, res) {
+        if(!req.session.login){return res.status(401).json({error: 'Unauthorised'})}
         res = await userModel.getUser(app, req.session.userId, res);
         return res;
     },
 
     amendItem: async function (app, req, res) {
+        if(!req.session.login){return res.status(401).json({error: 'Unauthorised'})}
         let amendedUser = await userModel.createUserFromRequest(req);
         res = await userModel.updateUser(app, amendedUser, req.session.userId, res)
         return res;
     },
 
     deleteItem: async function (app, req, res) {
+        if(!req.session.login){return res.status(401).json({error: 'Unauthorised'})}
         res = await userModel.deleteUser(app, req.session.userId, res);
         return res;
     },
